refactor(admin-api): extract Users & Lines actions into a controller
Sixth slice of the admin/api.php readability refactor.

- Add UserAjaxController with 5 actions: line, line_activity, adjust_credits,
  reg_user, ticket. Logic ported faithfully; nested branches flattened;
  comments English.
- Add LineStateTrait for the enable/disable/ban/unban/kill sub-actions that
  were byte-identical across the line, mag and enigma blocks (they all act on
  a line's user_id). DeviceAjaxController now uses the trait instead of its own
  private copy; its now-unused imports were dropped. The trait carries
  @phpstan-require-extends plus abstract ok()/fail() stubs so both PHPStan and
  Intelephense resolve the never-returning helpers.
- process (process-monitor) and profile (transcode profile) were left in
  api.php — they are not user/line actions.
- Register the 5 actions via $router->api(...) and delete their blocks from
  api.php (3330 -> 3142 lines; 1843 removed across the six slices).

Verified: php -l, make gates, make phpstan (0), make cs.
