+++ src/Public/Views/admin/api.php
@@ -304,198 +304,10 @@ if (isset($_SESSION['hash'])) {
 				exit();
 			}
 		}
-		if (RequestManager::get('action') == 'line') {
-			if (Authorization::check('adv', 'edit_user')) {
-				$rUserID = intval(RequestManager::get('user_id'));
-				$rSub = RequestManager::get('sub');
-
-				if ($rSub == 'delete') {
-					LineService::deleteLineById($rUserID);
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'enable') {
-					$db->query('UPDATE `lines` SET `enabled` = 1 WHERE `id` = ?;', $rUserID);
-					LineService::updateLineSignal($rUserID);
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'disable') {
-					$db->query('UPDATE `lines` SET `enabled` = 0 WHERE `id` = ?;', $rUserID);
-					LineService::updateLineSignal($rUserID);
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'ban') {
-					$db->query('UPDATE `lines` SET `admin_enabled` = 0 WHERE `id` = ?;', $rUserID);
-					LineService::updateLineSignal($rUserID);
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'unban') {
-					$db->query('UPDATE `lines` SET `admin_enabled` = 1 WHERE `id` = ?;', $rUserID);
-					LineService::updateLineSignal($rUserID);
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'kill') {
-					if (SettingsManager::get('redis_handler')) {
-						foreach (ConnectionTracker::getRedisConnections($rUserID, null, null, true, false, false) as $rConnection) {
-							ConnectionTracker::closeConnection($rConnection);
-						}
-					} else {
-						$db->query('SELECT * FROM `lines_live` WHERE `user_id` = ?;', $rUserID);
-
-						foreach ($db->get_rows() as $rRow) {
-							ConnectionTracker::closeConnection($rRow);
-						}
-					}
-
-					echo json_encode(array('result' => true));
-
-					exit();
-				} else {
-					echo json_encode(array('result' => false));
-
-					exit();
-				}
-			} else {
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-		}
-		if (RequestManager::get('action') == 'line_activity') {
-			if (Authorization::check('adv', 'connection_logs')) {
-				$rSub = RequestManager::get('sub');
-
-				if ($rSub != 'kill') {
-					echo json_encode(array('result' => false));
-
-					exit();
-				}
-
-				ConnectionTracker::closeConnection(RequestManager::get('pid'));
-				echo json_encode(array('result' => true));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
 		if (RequestManager::get('action') == 'process') {
 			if (Authorization::check('adv', 'process_monitor')) {
 				ApiClient::systemRequest(RequestManager::get('server'), array('action' => 'kill_pid', 'pid' => intval(RequestManager::get('pid'))));
 				echo json_encode(array('result' => true));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'adjust_credits') {
-			if (Authorization::check('adv', 'edit_reguser')) {
-				$rUser = UserRepository::getRegisteredUserById(RequestManager::get('id'));
-
-				if ($rUser && is_numeric(RequestManager::get('credits'))) {
-					$rCredits = intval($rUser['credits']) + intval(RequestManager::get('credits'));
-
-					if (0 <= $rCredits) {
-						$db->query('UPDATE `users` SET `credits` = ? WHERE `id` = ?;', $rCredits, $rUser['id']);
-						$db->query('INSERT INTO `users_credits_logs`(`target_id`, `admin_id`, `amount`, `date`, `reason`) VALUES(?, ?, ?, ?, ?);', $rUser['id'], $rUserInfo['id'], RequestManager::get('credits'), time(), RequestManager::get('reason'));
-						echo json_encode(array('result' => true));
-
-						exit();
-					}
-
-					echo json_encode(array('result' => false));
-
-					exit();
-				}
-
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'reg_user') {
-			if (Authorization::check('adv', 'edit_reguser')) {
-				$rSub = RequestManager::get('sub');
-
-				if ($rSub == 'delete') {
-					UserService::deleteRegisteredUser(RequestManager::get('user_id'), false, false, $rUserInfo['id']);
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'enable') {
-					$db->query('UPDATE `users` SET `status` = 1 WHERE `id` = ?;', RequestManager::get('user_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'disable') {
-					$db->query('UPDATE `users` SET `status` = 0 WHERE `id` = ?;', RequestManager::get('user_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				echo json_encode(array('result' => false));
-
-				exit();
-			}
-
-			echo json_encode(array('result' => false));
-
-			exit();
-		}
-		if (RequestManager::get('action') == 'ticket') {
-			if (Authorization::check('adv', 'ticket')) {
-				$rSub = RequestManager::get('sub');
-
-				if ($rSub == 'delete') {
-					TicketRepository::deleteById(RequestManager::get('ticket_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'close') {
-					$db->query('UPDATE `tickets` SET `status` = 0 WHERE `id` = ?;', RequestManager::get('ticket_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				if ($rSub == 'reopen') {
-					$db->query('UPDATE `tickets` SET `status` = 1 WHERE `id` = ?;', RequestManager::get('ticket_id'));
-					echo json_encode(array('result' => true));
-
-					exit();
-				}
-
-				echo json_encode(array('result' => false));
 
 				exit();
 			}
